Put Mesh class into separate file
